Move proto desc to framework
